{"schema":"decodex/commit/1","summary":"Move session bootstrap construction to owner","authority":"manual"}
